requirements.txt
Flask==2.3.1
Flask-Cors==3.0.10
keras==3.1.1
numpy==1.24.3
Pillow==9.3.0
tensorflow==2.16.1
Werkzeug==2.3.0